% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.max.in.range.R
\name{calc.max.in.range}
\alias{calc.max.in.range}
\title{Calculate the x and y values in a model where y is maximum}
\usage{
calc.max.in.range(df, poly.degree = 2, compute.model = F, y.col = NULL,
  x.col = NULL, x.min = 30, x.max = 200)
}
\arguments{
\item{df}{dataframe or dataframe group}

\item{poly.degree}{degree of the polynomial model; default is 2. Only 2 and 3 are currently available.}

\item{compute.model}{logical; whether or not to fit the underlying model. Default is F. If T, x.col and y.col must be specified.}

\item{y.col}{y column to fit in the model; required if compute.model = T}

\item{x.col}{x column to fit in the model; required if compute.model = T}

\item{x.min}{Minimum value of the range over which the maximum is to be computed}

\item{x.max}{Maximum value of the range over which the maximum is to be computed}
}
\value{
Returns a one-row dataframe with all grouping variables and the (x,y) values corresponding to the value at which y is maximal in the range (xmin, xmax).
}
\description{
This function is designed to evaluate a polynomial model (either calculating it from scratch using general.poly.fit or with a dataframe that already contains the output of general.poly.fit) between two variables, specified by the user, and then returns the (x,y) value where y is maximal in a given range. Since the model is a polynomial function, there should only ever be one max y.
}
