% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_induction_sigmoid.R
\name{fit_induction_sigmoid}
\alias{fit_induction_sigmoid}
\title{Fit induction curve data with a sigmoid}
\usage{
fit_induction_sigmoid(df, start.list = c(a = 120, b = 0.05, c = 120, d = 1),
  x.cutoff = "none")
}
\arguments{
\item{df}{Dataframe or dataframe subset with columns "timepoint", "med.red", "shock.pH.f", and "treatment" (last two are grouping variables if a dataframe subset)}

\item{start.list}{Edit the starting fitting parameters; default is list(a=120, b=0.05, c=120, d=1) (keep this form but change numbers)}

\item{x.cutoff}{maximum timepoint to consider in fitting; default is "none" (fits on all timepoints)}
}
\value{
dataframe with columns "treatment", "shock.pH.f", "timepoint", and "med.red" for plotting
}
\description{
This function takes raw data from a calibration curve experiment (yeast expressing pHluorin, flow cytometry data) and does all processing steps to construct a calibration curve that maps fluorescence ratio to pH. Takes *raw* (untransformed) FITC and BV510 channels as inputs!
}
