% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mm.R
\name{fit_mm}
\alias{fit_mm}
\title{Fit a mixture of two Gaussians to groups of data}
\usage{
fit_mm(df, output = "classification")
}
\arguments{
\item{df}{Dataframe or dataframe subset with a column 'rel.red' containing data to be fit}

\item{ouput}{Options are "classification", "posterior", "full"; default is "classification"}
}
\value{
a classified dataset ("classification", default), a dataset with the posteriors added as columns ("posterior"), or a dataset with posteriors and information about the mixtures ("full")
}
\description{
Fit a mixture of two Gaussians to groups of data
}
