% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_auc.R
\name{fitness_auc}
\alias{fitness_auc}
\title{Generate area under the curve for fitness data}
\usage{
fitness_auc(dat, normalize = FALSE)
}
\arguments{
\item{dat}{A dataframe or dataframe subset to act on}

\item{normalize}{Boolean; whether or not to normalize the data. Default is FALSE}
}
\value{
dataframe of area under the fitness curve
}
\description{
Data should be of the form x = recovery time and y = log ratio of stressed to unstressed.
}
