% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rolling_mean.R
\name{gen_rolling_mean}
\alias{gen_rolling_mean}
\title{Calculate rolling mean of a generic column in a dataframe (timeseries)}
\usage{
gen_rolling_mean(dat, data.column, time.column = "timepoint", k = "auto",
  k_val = 0.05)
}
\arguments{
\item{dat}{A dataframe or dataframe subset to act on.}

\item{data.column}{a string representing the name of the column containing the timeseries observations}

\item{time.column}{a string representign the name of the column containing the time; default is "timepoint"}

\item{k}{Indicates method to use for binwidth. If k is an integer, this will be the number of samples to include in the averaging bin. If k is "auto", the binwidth will be determined for each subset individually, size will be determined by k_val}

\item{k_val}{Only used if k = "auto"; the percentage of the total number of observations that will be averaged over for each subset}
}
\value{
a dataframe with columns "Time" and "fit" representing the rolling mean at each timepoint.
}
\description{
Takes a subset of a dataframe that represents a single experiment as input and calculates a rolling mean on the data with binwidth k.
}
