% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.poly.fit.R
\name{general.poly.fit}
\alias{general.poly.fit}
\title{General polynomial fitting function}
\usage{
general.poly.fit(df, x.col, y.col, poly.degree = 2)
}
\arguments{
\item{df}{dataframe or dataframe group}

\item{x.col}{name of the x column to be fit; must be a string}

\item{y.col}{name of the y column to be fit; must be a string}

\item{poly.degree}{optional; use to specify the degree of the polynomial model. Default is 2. Available options are 2 or 3.}
}
\value{
Returns a one-row dataframe with all grouping variables, and the parameters of the polynomial model as columns of the dataframe.
}
\description{
Operates on a dataframe or dataframe subset and does a polynomial fit between two of the columns.
}
