% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_log2_labels.R
\name{return_log2_labels}
\alias{return_log2_labels}
\title{Helper function to make log2 labels for axes}
\usage{
return_log2_labels(expo)
}
\arguments{
\item{expo}{A list of exponents; can be a numeric or character vector}
}
\value{
A list of expressions of the form 2^
}
\description{
Helper function to make log2 labels for axes
}
