% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_loglog.R
\name{scale_loglog}
\alias{scale_loglog}
\title{Scientific log-log scale}
\usage{
scale_loglog(...)
}
\description{
Creates a nice log 10 scale for plots (goes with scales package and ggplot2).
}
\examples{
ggplot(dat, aes(...)) + scale_loglog()
}
