% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_loglog2.R
\name{scale_loglog2}
\alias{scale_loglog2}
\title{Scientific log2 scale}
\usage{
scale_loglog2(...)
}
\description{
Creates a nice log 2 scale for plots (goes with scales package and ggplot2).
}
\examples{
ggplot(dat, aes(...)) + scale_x_loglog2()
}
