% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_log10nice.R
\name{scale_x_log10nice}
\alias{scale_x_log10nice}
\title{Scientific log scale for x}
\usage{
scale_x_log10nice(name = waiver(), omag = seq(-10, 20), ...)
}
\description{
Creates a nice log 10 scale for plots (goes with scales package and ggplot2).
}
\examples{
ggplot(dat, aes(...)) + scale_x_log10nice()
}
