% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_y_log2nice.R
\name{scale_y_log2nice}
\alias{scale_y_log2nice}
\title{Scientific log2 scale for y}
\usage{
scale_y_log2nice(name = NULL, omag = seq(-5, 5), ...)
}
\description{
Creates a nice log 2 scale for plots (goes with scales package and ggplot2).
}
\examples{
ggplot(dat, aes(...)) + scale_y_log2nice()
}
