% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scientific_10.R
\name{scientific_10}
\alias{scientific_10}
\title{Scientific 10}
\usage{
scientific_10(x)
}
\description{
This is a helper function used to construct nice scales for plots (goes with scales package and ggplot2).
}
