% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid}
\alias{sigmoid}
\title{pHluorin Calibration Curve Analysis}
\usage{
sigmoid(params, x)
}
\arguments{
\item{params}{character vector of parameters c(a, b, c, d); a is the max value, b is the steepness, c is the location of the inflection point, and d is the min value}

\item{x}{x values to generate the sigmoid for}
}
\value{
vector of sigmoid function values for each x point given
}
\description{
This function takes raw data from a calibration curve experiment (yeast expressing pHluorin, flow cytometry data) and does all processing steps to construct a calibration curve that maps fluorescence ratio to pH. Takes *raw* (untransformed) FITC and BV510 channels as inputs!
}
