% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ct_stresspH.R
\name{theme_ct_stresspH}
\alias{theme_ct_stresspH}
\title{Paper plotting theme}
\usage{
theme_ct_stresspH(base_size = 10)
}
\arguments{
\item{base_size}{text size; default is 10}
}
\description{
Adding this function to a ggplot changes the theme
}
