% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFasta.R
\name{writeFasta}
\alias{writeFasta}
\title{Write fasta from dataframe}
\usage{
writeFasta(data, filename)
}
\arguments{
\item{data}{dataframe containing two columns: ORF and seq}

\item{filename}{file to write fasta to}
}
\value{
returns the number of fasta files written
}
\description{
Adapated Nicholas Hathaway (https://bootstrappers.umassmed.edu/guides/main/r_writeFasta.html)
Function to write a dataframe of sequences to a fasta
}
\examples{
writeFasta(df \%>\% select(ORF,seq),filename);
}
